% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_id.R
\name{print.group_id_result}
\alias{print.group_id_result}
\title{Print method for group_id_result objects}
\usage{
\method{print}{group_id_result}(x, ...)
}
\arguments{
\item{x}{A group_id_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for group_id_result objects
}
